cleanup to utils
